Module Graphv_gles2_native.BlendFactor

Blending modes used for compositing. The current blend mode determines how a fill or stroke is composited to the current framebuffer. See here for more information on the different blending modes.

type t =

| Zero

| One

| Src_color

| One_minus_src_color

| Dst_color

| One_minus_dst_color

| Src_alpha

| One_minus_src_alpha

| Dst_alpha

| One_minus_dst_alpha

| Src_alpha_saturate