Module Buffer.Float

type t = (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val set : t -> int -> float -> unit

set t off value will set the value at offset off in the buffer.

val get : t -> int -> float

get t off will return the value at offset off.

val length : t -> int

Return the number of elements in the buffer.

val create : int -> t

Creates a new buffer. The buffer is initialized with zeros.

val blit : src:t -> s_off:int -> dst:t -> d_off:int -> len:int -> unit

Copy one buffer into another.

parameter src

The buffer to copy from.

parameter s_off

The offset to start copying from the src buffer.

parameter dst

The buffer to copy to.

parameter d_off

The offset to start copying to the dst buffer.

parameter len

The number of values to copy.

val fill : t -> float -> unit

Fill the buffer with the given value.