Module Buffer.UByte

type t = (int, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val set : t -> int -> int -> unit

set t off value will set the value at offset off in the buffer.

val get : t -> int -> int

get t off will return the value at offset off.

val length : t -> int

Return the number of elements in the buffer.

val sub : t -> int -> int -> t

sub t off len will create a sub array starting at offset off with length len. This operation does not copy any array values.

val create : int -> t

Creates a new buffer. The buffer is initialized with zeros.

val empty : t

The empty buffer. Can be used as a placeholder value. Has size 0.